Fix heap corruption caused by dangling C pointer
In func Simple, the inputBuffer and outputBuffer are created as slices, then
the address of each is copied to a C pointer (member of struct allocated with
malloc). At this point, the Go gc can free or move the memory - go only
guarantees that the gc won't free an object if the address of the object is
passed as a cgo function parameter, no such guarantee is made for pointers in
structs allocated with malloc.

Fix this by passing inputBuffer and outputBuffer as parameters to the cgo call.

Fixes:
 runtime: pointer 0xc000c32000 to unused region of span span.base()=0xc00062e000 span.limit=0xc00062fe00 span.state=1
 fatal error: found bad pointer in Go heap (incorrect use of unsafe or cgo?)